Fix group chats
It was impossible to add an item in a group chat with
Privacy mode turned on: https://core.telegram.org/bots/#privacy-mode

Now we force user to add an item as an argument in a group chat.
It's still possible to add an item in multiple steps in private chats.

This behaviour needs to be inproved for group chats,
because  it's hard to use command suggestions: you have to type
whole command by hands.