Improve group chats UX
In group chats, it's not necessary to use the /add command
with argumets now: an user can follow the same step-by-step process
as in the private chats. The only difference is that in a group
chat user will be forced to reply to bot's message. No additional
actions required from the user