Incapsulates integration with tgbotapi
We don't want to expose the intergration with tgbotapi to the code outside
the telegram package, so the package defines the APIClient type
that provides some proxy methods to tgbotapi.BotAPI.

Also this commit introduces some internal interfaces
which should simplify unit testing.

Internal interfaces are likely to change in further steps of the refactoring.